rename templatetag; fix image og html quotes
