+++ core/templates/core/opengraph.html
@@ -1,5 +1,5 @@
 {% load static %}
-{% load custom %}
+{% load opengraph %}
 
 <!-- HTML Meta Tags -->
 <title>{% meta_title %}</title>
@@ -10,7 +10,7 @@
 <meta property="og:type" content="website">
 <meta property="og:title" content="{% meta_title %}">
 <meta property="og:description" content="{% meta_description %}">
-<meta property="og:image" content={% static 'core/images/NYC-Noise-facebook-preview-TOO-BIG.png' %}>
+<meta property="og:image" content="{% static 'core/images/NYC-Noise-facebook-preview-TOO-BIG.png' %}">
 
 <!-- Twitter Meta Tags -->
 <meta name="twitter:card" content="summary_large_image">
@@ -18,4 +18,4 @@
 <meta property="twitter:url" content="{% meta_url %}">
 <meta name="twitter:title" content="{% meta_title %}">
 <meta name="twitter:description" content="{% meta_description %}">
-<meta name="twitter:image" content={% static 'core/images/NYC-Noise-facebook-preview-TOO-BIG.png' %}>
+<meta name="twitter:image" content="{% static 'core/images/NYC-Noise-facebook-preview-TOO-BIG.png' %}">
